+++ src/lib/logs.js
@@ -1,9 +1,13 @@
 import run from './run';
 
-export default function getLogs({ from, until = 'HEAD', cwd = undefined } = {}) {
+export default function getLogs({ from, until = 'HEAD', cwd = undefined, dir } = {}) {
   const range = from ? [from, until] : [until];
 
-  return run('git', ['log', range.join('..'), '--format=NEXT%n%H%n%aD%n%B'], { cwd })
+  return run(
+    'git',
+    ['log', range.join('..'), '--format=NEXT%n%H%n%aD%n%B', ...(dir ? ['--', dir] : [])],
+    { cwd }
+  )
     .then(({ stdout }) => stdout.split('NEXT'))
     .then(rawCommits =>
       rawCommits.slice(1).map(raw => {