+++ CHANGELOG.md
@@ -1,3 +1,36 @@
+<a name="1.0.0-beta.6"></a>
+# 1.0.0-beta.6 (2020-09-19)
+
+
+### Bug Fixes
+
+* **release:** Fix typo that prevents tag creation ([c8e215b](https://github.com/ls-age/bump-version/commits/c8e215b))
+* **release:** Skip tag creation in dry run only ([2b7a3e0](https://github.com/ls-age/bump-version/commits/2b7a3e0))
+* Bump version in monorepo package dir ([9731f42](https://github.com/ls-age/bump-version/commits/9731f42))
+* Compare correct tags ([a446c60](https://github.com/ls-age/bump-version/commits/a446c60))
+* Do not verify release files ([2b2568d](https://github.com/ls-age/bump-version/commits/2b2568d))
+* Fix typo ([#48](https://github.com/ls-age/bump-version/issues/48)) ([d1a7b76](https://github.com/ls-age/bump-version/commits/d1a7b76))
+* Prevent git checks for pnpm ([4390bcb](https://github.com/ls-age/bump-version/commits/4390bcb))
+* Prevent type error while getting tags ([48c852f](https://github.com/ls-age/bump-version/commits/48c852f))
+
+
+### Features
+
+* Add 'in-monorepo' command ([820fac7](https://github.com/ls-age/bump-version/commits/820fac7))
+* Basic monorepo support ([#45](https://github.com/ls-age/bump-version/issues/45)) ([a186499](https://github.com/ls-age/bump-version/commits/a186499))
+* Preview releases with the '--dry-run' option ([3348f2c](https://github.com/ls-age/bump-version/commits/3348f2c))
+* Run for subpackages with the '--dir' option ([f1d29a9](https://github.com/ls-age/bump-version/commits/f1d29a9))
+* Support publishing with pnpm ([#47](https://github.com/ls-age/bump-version/issues/47)) ([cc16bbc](https://github.com/ls-age/bump-version/commits/cc16bbc))
+* Support releasing packages in subdirs ([7e11b5a](https://github.com/ls-age/bump-version/commits/7e11b5a))
+
+
+### BREAKING CHANGES
+
+* release now returns an object
+
+
+
+
 <a name="0.2.1"></a>
 ## 0.2.1 (2019-10-29)